+++ src/masternodeman.h
@@ -158,4 +158,6 @@ public:
     void UpdateMasternodeList(CMasternodeBroadcast mnb);
 };
 
+void ThreadCheckObfuScationPool();
+
 #endif
