+++ src/obfuscation.h
@@ -352,6 +352,5 @@ public:
     void RelayCompletedTransaction(const int sessionID, const bool error, const int errorID);
 };
 
-void ThreadCheckObfuScationPool();
 
 #endif
